All fixed merged discrepencies and new changes related to Default instruction page
